Iterator changed to IntoIterator
`IntoIterator` is a more generic version of an iterator, that also
happens to support conversion of `CloneOnce` into the inner iterator
*before* iteration. This way we avoid taking the value out and putting
it back at each call of `next()` and also avoid using `RefCell`.

However, this change is formally backwards-incompatible because
`CloneOnce` ceased to implement `Iterator`, so any code that calls its
methods would break. For the intent of this crate, this should be
unlikely, so upgrading will be trivial in most cases.

Closes #1